src/Comrade/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.3
    # via discord.py
async-lru==1.0.2
    # via -r requirements.in
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
beautifulsoup4==4.9.3
    # via bs4
bs4==0.0.1
    # via
    #   -r requirements.in
    #   pydictionary
cachetools==4.2.1
    # via flask-discord
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via pynacl
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via
    #   flask
    #   pydictionary
cycler==0.10.0
    # via matplotlib
discord-pretty-help==1.2.1
    # via -r requirements.in
discord-py-slash-command==1.0.9.3
    # via -r requirements.in
discord.py==1.6.0
    # via
    #   -r requirements.in
    #   flask-discord
dnspython==2.1.0
    # via -r requirements.in
flask-discord==0.1.64
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-discord
futures==3.1.1
    # via goslate
fuzzywuzzy==0.18.0
    # via -r requirements.in
goslate==1.5.1
    # via pydictionary
idna==2.10
    # via
    #   requests
    #   yarl
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
kiwisolver==1.3.1
    # via matplotlib
markupsafe==1.1.1
    # via jinja2
matplotlib==3.3.4
    # via -r requirements.in
mpmath==1.2.1
    # via sympy
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.20.1
    # via
    #   -r requirements.in
    #   matplotlib
oauthlib==3.1.0
    # via
    #   flask-discord
    #   requests-oauthlib
pillow==8.1.0
    # via
    #   -r requirements.in
    #   matplotlib
pycparser==2.20
    # via cffi
pydictionary==2.0.1
    # via -r requirements.in
pyjwt==2.0.1
    # via flask-discord
pymongo==3.11.3
    # via -r requirements.in
pynacl==1.4.0
    # via -r requirements.in
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via matplotlib
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2021.1
    # via -r requirements.in
requests-oauthlib==1.3.0
    # via flask-discord
requests==2.25.1
    # via
    #   flask-discord
    #   pydictionary
    #   requests-oauthlib
six==1.15.0
    # via
    #   cycler
    #   pynacl
    #   python-dateutil
soupsieve==2.2
    # via beautifulsoup4
sympy==1.7.1
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via aiohttp
unidecode==1.2.0
    # via -r requirements.in
urllib3==1.26.3
    # via requests
werkzeug==1.0.1
    # via flask
yarl==1.6.3
    # via aiohttp
youtube-dl==2021.2.10
    # via -r requirements.in
